refactor: split player protocol module
